211110 일기 작성 및 일기 개별 조회 구현 ver.1
- 일기 작성 api 구현
- 일기 다이어리 id별로 조회 api 구현
- userFriendStatus 관련 클래스 생성
- 필요없는 home.jsp 관련 파일 제거
- 암호화 handler 생성 -> to be updated

+++ src/main/java/com/diary/frienda/request/Stickers.java
@@ -1,23 +1,23 @@
 package com.diary.frienda.request;
 
 public class Stickers {
-	private int locationX;
-	private int locationY;
+	private double locationX;
+	private double locationY;
 	private int sticker_id;
 	
-	public int getLocationX() {
+	public double getLocationX() {
 		return locationX;
 	}
-	public void setLocationX(int locationX) {
+	public void setLocationX(double locationX) {
 		this.locationX = locationX;
 	}
-	public int getLocationY() {
+	public double getLocationY() {
 		return locationY;
 	}
-	public void setLocationY(int locationY) {
+	public void setLocationY(double locationY) {
 		this.locationY = locationY;
 	}
-	public int getSticker_id() {
+	public double getSticker_id() {
 		return sticker_id;
 	}
 	public void setSticker_id(int sticker_id) {
